+++ packages/lumx-vue/src/components/select-button/index.ts
@@ -7,7 +7,11 @@ export {
     type MultipleSelectButtonProps,
     type SelectButtonButtonSlotProps,
 } from './SelectButton';
-export type { SelectTextFieldStatus, SelectButtonTranslations } from '@lumx/core/js/utils/select/types';
+export type {
+    SelectListStatus,
+    SelectListStatus as SelectButtonStatus,
+    SelectButtonTranslations,
+} from '@lumx/core/js/utils/select/types';
 
 /** Selectable option within the dropdown list. */
 export const SelectButtonOption = _ComboboxOption;